% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_jmp_plus.R
\name{choosemodel}
\alias{choosemodel}
\title{Test poolability of model}
\usage{
choosemodel(data, outcome, poolMSE, debug)
}
\arguments{
\item{data}{A dataframe containing variables "time" and "batch" to be analysed.}

\item{outcome}{Outcome variable name in quotation marks.}

\item{poolMSE}{Use pooled MSE? To turn on, poolMSE = 1.}

\item{debug}{Debugging. To turn on, debug = 1.}
}
\value{
A list of (i. Chosen model as an integer from 1-4, ii. String description of chosen model).
}
\description{
This uses the Type 1 Sum of Squares ANOVA results from the time*batch interaction model
}
\examples{
  # Choose model to use by p values and whether MSE is to be pooled or not
  # model <- choosemodel(data = data, outcome = "result_value", poolMSE = 0, debug = 0)
}
