% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eberhart_Russel_anova.R
\name{Eberhart_Russel_anova}
\alias{Eberhart_Russel_anova}
\title{ANOVA of Eberhart & Russel’s Model}
\usage{
Eberhart_Russel_anova(data, Rep, G, E, Y)
}
\arguments{
\item{data}{data.frame}

\item{Rep}{Replication Factor}

\item{G}{Genotypes Factor}

\item{E}{Environment Factor}

\item{Y}{Response Variable}
}
\value{
Additive ANOVA
}
\description{
ANOVA of Eberhart & Russel’s Model
}
\examples{
data(ge_data)
Yield.ERANOVA <-
     Eberhart_Russel_anova(
          data = ge_data
       , G     = Gen
       , E     = Env
       , Y     = Yield
       )[[1]]
Yield.ERANOVA

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
