% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{CSL312 sFP 36 months stability data}
\format{
\subsection{\code{example_data}}{

A data frame with 396 rows and 16 columns:
\describe{
\item{Time (months)}{Stability time point}
\item{Temperature}{Storage temperature}
\item{Lot}{Year}
\item{Quality Attribute}{Year}
\item{Lower Limit}{Lower limit of quality attribute}
\item{Upper Limit}{Upper limit of quality attribute}
\item{Units}{Unit of quality attribute}
\item{Result Value}{Year}

}
}
}
\source{
CSL312 sFP 36 months stability data
}
\usage{
example_data
}
\description{
Stability data for CSL312 sFP up to 36 months with 3 batches
}
\keyword{datasets}
