% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_data.R
\docType{data}
\name{ge_data}
\alias{ge_data}
\title{Data for Genotypes by Environment Interaction (GEI)}
\format{A \code{data.frame} 1320 obs. of  6 variables.}
\usage{
data(ge_data)
}
\description{
\code{ge_data} is used for performing Genotypes by Environment Interaction (GEI) Analysis.
}
\details{
\itemize{
       \item{Gen} Genotype
       \item{Institute} Institute
       \item{Rep} Replicate
       \item{Block} Block
       \item{Env} Environment
       \item{Yield} Yield Response
       }
}
\examples{
data(ge_data)

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
