% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_effects.R
\name{ge_effects}
\alias{ge_effects}
\title{Genotype by Environment Interaction Effects}
\usage{
ge_effects(data, G, E, Y)
}
\arguments{
\item{data}{data.frame}

\item{G}{Genotypes Factor}

\item{E}{Environment Factor}

\item{Y}{Response Variable}
}
\value{
Effects
}
\description{
Provides Genotype by Environment Interaction Effects
}
\examples{

data(ge_data)
Yield.Effects <- ge_effects(data = ge_data, G = Gen, E = Env, Y = Yield)
names(Yield.Effects)

Yield.Effects$ge_means
Yield.Effects$ge_effects
Yield.Effects$gge_effects
}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
