% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_means.R
\name{ge_means}
\alias{ge_means}
\title{Genotype by Environment Interaction Means and Ranks}
\usage{
ge_means(data, G, E, Y)
}
\arguments{
\item{data}{data.frame}

\item{G}{Genotypes Factor}

\item{E}{Environment Factor}

\item{Y}{Response Variable}
}
\value{
Means and Ranks
}
\description{
Provides Genotype by Environment Interaction Means along with their Ranks
}
\examples{

data(ge_data)

Yield.ge_means <- ge_means(data = ge_data, G = Gen, E = Env, Y = Yield)

Yield.ge_means$ge_means1
Yield.ge_means$ge_means
Yield.ge_means$ge_ranks

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
