% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_jmp_plus.R
\name{models}
\alias{models}
\title{Get all models and results}
\usage{
models(data, outcome, colours_named, debug)
}
\arguments{
\item{data}{A dataframe containing variables "time" and "batch" to be analysed.}

\item{outcome}{Outcome variable name in quotation marks.}

\item{colours_named}{A list of named colours to use for plotting each batch.}

\item{debug}{Debugging. To turn on, debug = 1.}
}
\value{
A list of (i. Model coefficients, ii. 90\% and 95\% confidence intervals, iii. Model diagnostics - residuals and fitted values).
}
\description{
Model numbering following JMP output. SAS Macro uses reverse numbering
}
\examples{
  # Choose model to use by p values and whether MSE is to be pooled or not
  # model <- choosemodel(data = data, outcome = "result_value", poolMSE = 0, debug = 0)
}
