% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_jmp_plus.R
\name{plot}
\alias{plot}
\title{Plot results from chosen model}
\usage{
plot(
  data,
  outcome,
  label,
  lowerlimit,
  upperlimit,
  chosenmodel,
  confRes,
  expiryRes,
  backtransf,
  colours_named
)
}
\arguments{
\item{data}{A result dataset of observed quality attribute values with variables "time" (in months or transformed) and "batch".}

\item{outcome}{Quality attribute outcome variable name in quotation marks.}

\item{label}{y-axis label.}

\item{lowerlimit}{Lower limit to plot on graph.}

\item{upperlimit}{Upper limit to plot on graph.}

\item{chosenmodel}{A numeric integer (1-4) of the chosen model from choosemodel()[[1]].}

\item{confRes}{A result dataset of predicted values from models()[[2]].}

\item{expiryRes}{A result dataset of expiry dates by model and batch from limits().}

\item{backtransf}{Formula for backtransformation of outcome variable. Supported formula strings are c("Time","log2","Time^2","Time^3","Time^4","Time^5") corresponding to the indicator formula for c(no transformation, Log 2 transformation, square root of time, cube root of time, fourth root of time, fifth root of time).}

\item{colours_named}{A list of named colours to use for plotting each batch.}
}
\value{
List of (i. Expiry date in months, ii. Plot of all observed values with estimated model mean and one-/two-sided 95\%CI)
}
\description{
Variables should be named "time" and "batch"
}
\examples{
    # Choose model to use by p values and whether MSE is to be pooled or not
    # model <- choosemodel(data = data, outcome = "result_value", poolMSE = 0, debug = 0)

    # Set colours using all batches in dataset
    # colours <- c("#FC1921","#0E56A5","#F5C017","#975DA2","#00A28A",
    #              "#DA2877","#F06125","#03B3BE","#C6D92D","#572A7B")
    # colours_named <- setNames(object = colours,
    #                           nm = c(unique(data_transf$batch),"Common Intercept"))

    # Save confidence interval results from all models
    # results <- models(data = data, outcome = "result_value",
    #                   colours_named = colours_named, debug = 0)

    # Get expiry date of all batches from all models by number of limits provided
    # exp <- limits(data = results[[2]], lowerlimit = 90, upperlimit = 110, debug = 0)

    # Plot results from chosen model
    # expiryplot <- plot(data = data, outcome = "result_value", label = "Potency (\%)",
    #                    lowerlimit = lowerlimit, upperlimit = upperlimit,
    #                    chosenmodel = model[[1]], confRes = results[[2]],
    #                    expiryRes = exp, backtransf = backtransform, colours_named = colours_named)
}
