% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_jmp_plus.R
\name{qq_jmp}
\alias{qq_jmp}
\title{Test qq plot
Draw a QQ plot of residuals based on JMP formula}
\usage{
qq_jmp(df, variable)
}
\arguments{
\item{df}{A dataframe containing a continuous variable to graph qq plot.}

\item{variable}{Variable name to graph qq plot in quotation marks.}
}
\value{
A qq plot.
}
\description{
Test qq plot
Draw a QQ plot of residuals based on JMP formula
}
\examples{
  # Draw qq plot of residuals
  # m1_qq <- qq_jmp(df = m1_diag, variable = "residuals")
}
