% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_kang.R
\name{stab_kang}
\alias{stab_kang}
\alias{stab_kang.default}
\title{Additive ANOVA for Genotypes by Environment Interaction (GEI) model}
\usage{
stab_kang(.data, .y, .rep, .gen, .env)

\method{stab_kang}{default}(.data, .y, .rep, .gen, .env)
}
\arguments{
\item{.data}{data.frame}

\item{.y}{Response Variable}

\item{.rep}{Replication Factor}

\item{.gen}{Genotypes Factor}

\item{.env}{Environment Factor}
}
\value{
Additive ANOVA
}
\description{
Additive ANOVA for Genotypes by Environment Interaction (GEI) model
}
\examples{
data(ge_data)
YieldKang <-
     stab_kang(
            .data = ge_data
          , .y    = Yield
          , .rep  = Rep
          , .gen  = Gen
          , .env  = Env
      )
YieldKang

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
