% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_measures.R
\name{stab_measures}
\alias{stab_measures}
\title{Stability Measures for Genotypes by Environment Interaction (GEI)}
\usage{
stab_measures(data, G, E, Y)
}
\arguments{
\item{data}{data.frame}

\item{G}{Genotypes Factor}

\item{E}{Environment Factor}

\item{Y}{Response Variable}
}
\value{
Stability Measures
}
\description{
Stability Measures for Genotypes by Environment Interaction (GEI)
}
\examples{

data(ge_data)
Yield.StabMeasures <- stab_measures(data = ge_data, G = Gen, E = Env, Y = Yield)
Yield.StabMeasures

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
