% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab_reg.R
\name{stab_reg}
\alias{stab_reg}
\title{Individual Regression for each Genotype}
\usage{
stab_reg(data, Rep, G, E, Y)
}
\arguments{
\item{data}{data.frame}

\item{Rep}{Replication Factor}

\item{G}{Genotypes Factor}

\item{E}{Environment Factor}

\item{Y}{Response Variable}
}
\value{
Additive ANOVA
}
\description{
Individual Regression for each Genotype in Genotypes by Environment Interaction (GEI)
}
\examples{

data(ge_data)
Yield.StabReg <- stab_reg(data = ge_data, G = Gen, E = Env, Y = Yield)
Yield.StabReg

}
\references{
Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
