% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_jmp_plus.R
\name{stability}
\alias{stability}
\title{Packaging all helper functions into one function call}
\usage{
stability(
  data,
  outcome,
  label,
  lowerlimit,
  upperlimit,
  poolMSE,
  backtransform,
  colours_named,
  debug
)
}
\arguments{
\item{data}{A result dataset of observed quality attribute values with variables "time" (in months or transformed) and "batch".}

\item{outcome}{Quality attribute outcome variable name in quotation marks.}

\item{label}{y-axis label.}

\item{lowerlimit}{Lower limit to plot on graph.}

\item{upperlimit}{Upper limit to plot on graph.}

\item{poolMSE}{Use pooled MSE? To turn on, poolMSE = 1.}

\item{backtransform}{Formula for backtransformation of outcome variable. Supported formula strings are c("Time","log2","Time^2","Time^3","Time^4","Time^5") corresponding to the indicator formula for c(no transformation, Log 2 transformation, square root of time, cube root of time, fourth root of time, fifth root of time).}

\item{colours_named}{A list of named colours to use for plotting each batch.}

\item{debug}{Debugging. To turn on, debug = 1.}
}
\value{
List of (i. Expiry date in months, ii. Plot of all observed values with estimated model mean and one-/two-sided 95\%CI)
}
\description{
Packaging all helper functions into one function call
}
\examples{

    # Create dataset to be analysed in long format with time variable with
    # transformed time where necessary and corresponding backtransform formula variable

    # data_transf <- data \%>\%
    #   mutate(time = case_when(quality_attribute == "SE HPLC Monomer" ~ sqrt(time),
    #                           quality_attribute == "SE HPLC HMWS" ~ sqrt(time),
    #                           quality_attribute == "CE reducing" ~ sqrt(time),
    #                           quality_attribute == "CE non reducing LMWS" ~ sqrt(time),
    #                           TRUE ~ time),
    #           backtransform = case_when(quality_attribute == "SE HPLC Monomer" ~ "Time^2",
    #                                     quality_attribute == "SE HPLC HMWS" ~ "Time^2",
    #                                     quality_attribute == "CE reducing" ~ "Time^2",
    #                                     quality_attribute == "CE non reducing LMWS" ~ "Time^2",
    #                                     TRUE ~ "Time"))

    # Set colours using all batches in dataset
    # colours <- c("#FC1921","#0E56A5","#F5C017","#975DA2","#00A28A",
    #              "#DA2877","#F06125","#03B3BE","#C6D92D","#572A7B")
    # colours_named <- setNames(object = colours,
    #                           nm = c(unique(data_transf$batch),"Common Intercept"))

    # Get vectors of quality attributes, labels, lower limits and upper limits
    # qa <- data_dist$quality_attribute
    # lbl <- data_dist$label
    # ll <- data_dist$lower_limit
    # ul <- data_dist$upper_limit

    # Initialise empty results list
    # stab <- vector(mode = "list", length = length(qa))

    # for (i in 1:1) {

       # Get subset of data containing only one quality attribute
       # data_qa <- subset(data_transf, quality_attribute == qa[i])

       # Get backtransform formula string
       # bt <- unique(data_qa$backtransform)

       # Call stability function and save results to list
       # stab[[i]] <- stability(data = data_qa, outcome = "result_value", label = lbl[i],
       #                        lowerlimit = ll[i], upperlimit = ul[i], poolMSE = 0,
       #                        backtransform = bt, colours_named = colours_named, debug = 0)
    # }
}
