% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{stability}
\alias{stability}
\title{Evaluation of model stability under distribution shift}
\usage{
stability(model, param, E = NULL)
}
\arguments{
\item{model}{An object as returned by lm or glm}

\item{param}{A coefficient in summary(model)}

\item{E}{A subset of predictors in the formula of the model}
}
\description{
Evaluation of model stability under distribution shift
}
\examples{
fit <- lm(sr ~ pop15 + pop75 + dpi + ddpi,
data = LifeCycleSavings)
stability(fit,param="pop15",E="dpi")
stability(fit,param="pop15")

utils::data(anorexia, package = "MASS")
fit <- glm(Postwt ~ Prewt + Treat,family = gaussian, data = anorexia)
stability(fit,param = "TreatCont")

}
